requirements.txt
streamlit
torch
torchvision
timm
# Remove opencv-python as it's causing deployment issues
pillow
scikit-learn
matplotlib
seaborn
numpy
scipy